+++ packages/core/src/controllers/ModalController.ts
@@ -60,6 +60,7 @@ export const ModalController = {
     } else {
       RouterController.reset('Connect')
     }
+    console.log('open', state.open)
     state.open = true
     PublicStateController.set({ open: true })
     EventsController.sendEvent({
